feature: migration & db setup
Add config section "migrations" to setup the migration manager and run
all pending migrations on startup when configured to do so using thw
"runOnStartup" setting.

Add "createDb" setting.
When "true" and the configured database does
not exist, the plugin tires to create the databse using the credentials
confgured in the server section.
When "false" and the database does not exist an error if thown on
start-up